Move Schema to its now file to extract from fields.py but avoid circular imports
