+++ pydantic/schema.py
@@ -8,10 +8,12 @@ from uuid import UUID
 from . import main
 from .fields import Field, Shape
 from .json import pydantic_encoder
+from .schema_class import Schema
 from .types import DSN, UUID1, UUID3, UUID4, UUID5, DirectoryPath, EmailStr, FilePath, Json, NameEmail, UrlStr
 from .utils import clean_docstring, lenient_issubclass
 
 __all__ = [
+    'Schema',
     'schema',
     'model_schema',
     'field_schema',
